Complete Overhaul
Removed the Listener script. Created only one input listener in the macroManager script. Converted the ToolPopup function to a class, the class interacts with the main gui using callbacks. There is a new on_submit, and on_cancel call back. Fixed bug when you edited duplicated macros, and if you edited the hokey it would apply that hotkey to all the hotkeys that were duplicated.